Invoke Code Climate in its own shell script
We are adding Windows as a third OS to Travis CI, and the Code Climate
reporter cannot run on Windows. With this extra consideration, the
logic gets so complex that keeping it all in .travis.yml would make
that file messy. The solution recommended by Travis is to spin it off
into a separate shell script.

The shell script is formatted with shfmt and checked with shellcheck.